Monotone-Parent: 7e7946ac70e135f3e208b210455b3238dd36d951 Monotone-Revision: daae15bcf9e4793ea2109c4e41219071a3e94636
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2010-08-30T19:49:58
Monotone-Branch: ca.inverse.sogo

+++ SOPE/NGCards/iCalPerson.m
@@ -247,8 +247,8 @@
 
 /* comparison */
 
-- (unsigned)hash {
-  if([self email])
+- (NSUInteger) hash {
+  if ([self email])
     return [[self email] hash];
   return [super hash];
 }
